DAO Testing. All examples complete
- Changed parameter of getLabTest(Encounter) to getLabTest(Order) in Service class
- Changed getByUUid methods to use Criteria instead of Query in DAO
- Changed Restrictions in DAO to match Ids instead of whole objects
- Changed sortWeight attribute type from Integer to Double in LabTestAttributeType, as was desired
- Fixed the length of datatype_config and handler_config in liquibase
- More objects added in XML dataset
- Changed order IDs in XML dataset to avoid conflict from existing test set

+++ api/src/main/java/org/openmrs/module/commonlabtest/LabTestAttributeType.java
@@ -49,7 +49,7 @@ public class LabTestAttributeType extends BaseAttributeType<LabTest> implements 
 	
 	@Field
 	@Column(name = "sort_weight")
-	private Integer sortWeight;
+	private Double sortWeight;
 	
 	public LabTestAttributeType() {
 	}
@@ -76,11 +76,11 @@ public class LabTestAttributeType extends BaseAttributeType<LabTest> implements 
 		this.labTestType = labTestType;
 	}
 	
-	public Integer getSortWeight() {
+	public Double getSortWeight() {
 		return sortWeight;
 	}
 	
-	public void setSortWeight(Integer sortWeight) {
+	public void setSortWeight(Double sortWeight) {
 		this.sortWeight = sortWeight;
 	}
 }
